Prototype fixes.
[git-p4: depot-paths = "//depot/openwatcom/": change = 24642]

+++ bld/browser/cpp/module.cpp
@@ -284,7 +284,7 @@ bool Module::findRefSyms( WVList *list, Symbol * sym )
  * these are the routines which are called by the debug reading library
 */
 
-static void dbgReadHook( void * obj, dr_section sect, void * buf, int len )
+static void dbgReadHook( void * obj, dr_section sect, void * buf, size_t len )
 //-------------------------------------------------------------------------
 {
     ((Module *)obj)->readSect( sect, buf, len );
