+++ bld/mstools/c/fuzzy.c
@@ -173,7 +173,7 @@ static unsigned hash_symbol_name( const void *symbol )
 /*
  * Used by ORL.
  */
-static void *obj_read( void *hdl, int len )
+static void *obj_read( void *hdl, size_t len )
 /*****************************************/
 {
     ListElem *          newelem;
